LayoutTests/fast/dom/offset-position-writing-modes.html

<script src="../../fast/js/resources/js-test-pre.js"></script>
<div style="position:relative">
<div style="-webkit-writing-mode: vertical-rl; background-color: green; width: 100px; height: 100px;">
    <div style="width: 5px"></div>
    <div id=vertical style="background-color: yellow; width: 10px; position: relative; left: -20px;"></div>
</div>
</div>

<div style="position:relative">
<div style="-webkit-writing-mode: horizontal-bt; background-color: red; width: 100px; height: 100px;">
    <div style="height: 5px"></div>
    <div id=horizontal style="background-color: orange; height: 10px; position: relative; top: -20px;"></div>
</div>
</div>

<pre id=console></pre>

<script>
shouldBe('document.getElementById("vertical").offsetLeft', '65');
shouldBe('document.getElementById("horizontal").offsetTop', '65');
</script>
<script src="../../fast/js/resources/js-test-post.js"></script>
