LayoutTests/fast/dom/offset-position-writing-modes-expected.txt

PASS document.getElementById("vertical").offsetLeft is 65
PASS document.getElementById("horizontal").offsetTop is 65
PASS successfullyParsed is true

TEST COMPLETE
